+++ src/screens/wallets.tsx
@@ -9,7 +9,7 @@ import {onBannerClaimAirdrop} from '@app/event-actions/on-banner-claim-airdrop';
 import {onBannerSnoozeUntil} from '@app/event-actions/on-banner-snooze-until';
 import {showModal} from '@app/helpers';
 import {useTypedNavigation, useWallets} from '@app/hooks';
-import {useBanner} from '@app/hooks/use-banner';
+import {useBanners} from '@app/hooks/use-banners';
 import {useWalletConnectSessions} from '@app/hooks/use-wallet-connect-sessions';
 import {I18N, getText} from '@app/i18n';
 import {BannerButton} from '@app/models/banner';
@@ -21,7 +21,7 @@ export const WalletsWrapper = () => {
   const wallets = useWallets();
   const [visibleRows, setVisibleRows] = useState(wallets.visible);
   const {activeSessions} = useWalletConnectSessions();
-  const banner = useBanner();
+  const banners = useBanners();
   const [walletConnectSessions, setWalletConnectSessions] = useState<
     SessionTypes.Struct[][]
   >([]);
@@ -161,7 +161,7 @@ export const WalletsWrapper = () => {
     <Wallets
       balance={balance}
       wallets={visibleRows}
-      banner={banner}
+      banners={banners}
       walletConnectSessions={walletConnectSessions}
       onWalletConnectPress={onWalletConnectPress}
       onPressSend={onPressSend}
